Update simple-register-login.cpp
Changed how user interacts with login and create session. User is free to choose and option to create, login or exit from the program. Creating a user will save username and password on a .txt document and can be read to directly initiate a login session when starting up the program after closing.